# 04-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinanceIntel Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions/": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Tx",
        "operationId": "create_tx_api_v1_transactions__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Tx",
        "operationId": "list_tx_api_v1_transactions__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Per Page"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get Tx",
        "operationId": "get_tx_api_v1_transactions__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Update Tx",
        "operationId": "update_tx_api_v1_transactions__id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete Tx",
        "operationId": "delete_tx_api_v1_transactions__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/pnl": {
      "get": {
        "tags": [
          "reports",
          "reports"
        ],
        "summary": "Pnl Report",
        "operationId": "pnl_report_api_v1_reports_pnl_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PnLResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/cash-flow": {
      "get": {
        "tags": [
          "reports",
          "reports"
        ],
        "summary": "Cash Flow Report",
        "operationId": "cash_flow_report_api_v1_reports_cash_flow_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "as_of_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/diagnostics/spend": {
      "get": {
        "tags": [
          "reports",
          "reports"
        ],
        "summary": "Spend Discrepancy Report",
        "operationId": "spend_discrepancy_report_api_v1_reports_diagnostics_spend_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendDiscrepancyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashFlowResult": {
        "properties": {
          "transaction_balance": {
            "type": "string",
            "format": "decimal",
            "title": "Transaction Balance"
          },
          "held_payouts": {
            "type": "string",
            "format": "decimal",
            "title": "Held Payouts"
          },
          "available_balance": {
            "type": "string",
            "format": "decimal",
            "title": "Available Balance"
          },
          "average_daily_spend_30d": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Daily Spend 30D"
          },
          "runway_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runway Days"
          }
        },
        "type": "object",
        "required": [
          "transaction_balance",
          "held_payouts",
          "available_balance",
          "average_daily_spend_30d",
          "runway_days"
        ],
        "title": "CashFlowResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PnLResult": {
        "properties": {
          "revenue": {
            "type": "string",
            "format": "decimal",
            "title": "Revenue"
          },
          "ad_spend": {
            "type": "string",
            "format": "decimal",
            "title": "Ad Spend"
          },
          "gross_profit": {
            "type": "string",
            "format": "decimal",
            "title": "Gross Profit"
          },
          "consumables": {
            "type": "string",
            "format": "decimal",
            "title": "Consumables"
          },
          "operating_expenses": {
            "type": "string",
            "format": "decimal",
            "title": "Operating Expenses"
          },
          "ebitda": {
            "type": "string",
            "format": "decimal",
            "title": "Ebitda"
          },
          "depreciation": {
            "type": "string",
            "format": "decimal",
            "title": "Depreciation"
          },
          "ebit": {
            "type": "string",
            "format": "decimal",
            "title": "Ebit"
          },
          "interest": {
            "type": "string",
            "format": "decimal",
            "title": "Interest"
          },
          "ebt": {
            "type": "string",
            "format": "decimal",
            "title": "Ebt"
          },
          "tax": {
            "type": "string",
            "format": "decimal",
            "title": "Tax"
          },
          "net_profit": {
            "type": "string",
            "format": "decimal",
            "title": "Net Profit"
          },
          "gross_margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Margin"
          },
          "ebitda_margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ebitda Margin"
          },
          "ebit_margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ebit Margin"
          },
          "ebt_margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ebt Margin"
          },
          "net_margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Margin"
          },
          "margin": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin"
          }
        },
        "type": "object",
        "required": [
          "revenue",
          "ad_spend",
          "gross_profit",
          "consumables",
          "operating_expenses",
          "ebitda",
          "depreciation",
          "ebit",
          "interest",
          "ebt",
          "tax",
          "net_profit",
          "gross_margin",
          "ebitda_margin",
          "ebit_margin",
          "ebt_margin",
          "net_margin",
          "margin"
        ],
        "title": "PnLResult"
      },
      "SpendDiscrepancyResult": {
        "properties": {
          "tracker_spend": {
            "type": "string",
            "format": "decimal",
            "title": "Tracker Spend"
          },
          "actual_spend": {
            "type": "string",
            "format": "decimal",
            "title": "Actual Spend"
          },
          "discrepancy": {
            "type": "string",
            "format": "decimal",
            "title": "Discrepancy"
          }
        },
        "type": "object",
        "required": [
          "tracker_spend",
          "actual_spend",
          "discrepancy"
        ],
        "title": "SpendDiscrepancyResult"
      },
...
      "TransactionCreate": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "amount": {
            "type": "string",
            "format": "decimal",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "type",
          "category",
          "amount",
          "currency",
          "occurred_on"
        ],
        "title": "TransactionCreate"
      },
      "TransactionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_amount": {
            "type": "string",
            "format": "decimal",
            "title": "Total Amount"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "total_amount",
          "page",
          "per_page"
        ],
        "title": "TransactionListResponse"
      },
      "TransactionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "amount": {
            "type": "string",
            "format": "decimal",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "category",
          "amount",
          "currency",
          "occurred_on"
        ],
        "title": "TransactionOut"
      },
      "TransactionUpdate": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "type": "object",
        "title": "TransactionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}